+++ CHANGES.rst
@@ -6,6 +6,10 @@ Changelog
 
 New Features:
 
+- Add user registration to @users POST
+  https://github.com/plone/plone.restapi/pull/255
+  [csenger]
+
 - Add @vocabularies endpoint.
   https://github.com/plone/plone.restapi/issues/212
   [timo,csenger,sneridagh]